requirements.txt
protobuf==3.20.0
google-ads==18.0.0
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
python-dateutil
pyyaml
streamlit
google-cloud
google-cloud-storage
